refactor: statically define information of VMs
The driver is used to test hypervisor implementation.
For now drop use of FFA_PARTITION_INFO_GET in the driver's
initalisation. All FF-A partitions are expected to have an
UUID, which was not enforced until FF-A v1.2 updates.

As such, statically define the hf_vms array. For the vcpus
structures malloc is still used.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I669da31ffebc7ad9938124c57c3cc2ea6ae53d06